add manager -_-
