+++ Manager/go.mod
@@ -1,0 +1,3 @@
+module Manager
+
+go 1.21.0
