+++ go.mod
@@ -1,4 +1,4 @@
-module HttpServer
+module httpServer
 
 go 1.21.1
